+++ generators/prerequisites/templates/base.rb
@@ -1,6 +1,6 @@
 # The base model. Models allow to communicate with
 # AUT (application under test) api endpoints. They are used by
-# FactoryGirl on create
+# FactoryBot on create
 #
 # To implement a custom model, override the following methods:
 # * {Base.find}